@aaronc 2023-05-23
> Can we change sdk.MustSortJson to use this then?

There should be no more need to MustSortJson. See #16062