@kocubinski 2023-05-30
@odeke-em This should fix integration tests. https://github.com/cosmos/cosmos-sdk/pull/16338 We just need to disable sorting in the legacy parity tests.  Strictly speaking, sorting JSON fields was a feature of tx signing not marshalling so this makes sense.